Changed views, likes and downloads fields in Photo and PhotoDTO from Int to Long type to prevent errors with older photos (Int overflow)

+++ app/src/main/java/com/andrii_a/walleria/data/remote/dto/photo/PhotoDTO.kt
@@ -13,9 +13,9 @@ data class PhotoDTO(
     @SerializedName("blur_hash")
     val blurHash: String?,
     val color: String? = "#E0E0E0",
-    val views: Int?,
-    val downloads: Int?,
-    val likes: Int?,
+    val views: Long?,
+    val downloads: Long?,
+    val likes: Long?,
     @SerializedName("liked_by_user")
     val likedByUser: Boolean,
     val description: String?,
